+++ lib/currency.ex
@@ -44,9 +44,17 @@ defmodule Cldr.Currency do
   @doc """
   Returns a `Currency` struct created from the arguments.
 
+  ## Options
+
   * `currency` is a custom currency code of a format defined in ISO4217
 
   * `options` is a map of options representing the optional elements of the `%Currency{}` struct
+
+  ## Returns
+
+  * `{:ok, Cldr.Currency.t}` or
+
+  * `{:error, {exception, message}}`
 
   ## Example
 
@@ -63,7 +71,7 @@ defmodule Cldr.Currency do
         tender: false}}
 
       iex> Cldr.Currency.new(:XBC)
-      {:error, "Currency :XBC is already defined"}
+      {:error, {Cldr.CurrencyAlreadyDefined, "Currency :XBC is already defined"}}
 
   """
   @spec new(binary | atom, map | list) :: t | {:error, binary}
@@ -84,16 +92,18 @@ defmodule Cldr.Currency do
   end
 
   def new(currency, options) when is_map(options) do
-    with false <- known_currency?(currency),
-         {:ok, currency_code} <- make_currency_code(currency)
+    with \
+      {:error, _currency} <- Cldr.validate_currency(currency),
+      {:ok, currency_code} <- make_currency_code(currency)
     do
-      options = @currency_defaults
-      |> Map.merge(options)
-      |> Map.merge(%{code: currency_code})
+      options =
+        @currency_defaults
+        |> Map.merge(options)
+        |> Map.merge(%{code: currency_code})
 
       {:ok, struct(__MODULE__, options)}
     else
-      true -> {:error, "Currency #{inspect currency} is already defined"}
+      {:ok, _} -> {:error, {Cldr.CurrencyAlreadyDefined, "Currency #{inspect currency} is already defined"}}
       error -> error
     end
   end
@@ -102,30 +112,38 @@ defmodule Cldr.Currency do
   Returns the appropriate currency display name for the `currency`, based
   on the plural rules in effect for the `locale`.
 
+  ## Options
+
   * `number` is an integer, float or `Decimal`
 
   * `currency` is any currency returned by `Cldr.Currency.known_currencies/0`
 
   * `options` is a keyword list of options
-    * `:locale` is any locale returned by `Cldr.Locale.new!/1`.  The
+    * `:locale` is any locale returned by `Cldr.Locale.new!/1`. The
     default is `Cldr.get_current_locale/0`
+
+  ## Returns
+
+  * `{:ok, plural_string}` or
+
+  * `{:error, {exception, message}}`
 
   ## Examples
 
       iex> Cldr.Currency.pluralize 1, :USD
-      "US dollar"
+      {:ok, "US dollar"}
 
       iex> Cldr.Currency.pluralize 3, :USD
-      "US dollars"
+      {:ok, "US dollars"}
 
-      iex> Cldr.Currency.pluralize 12, :USD, locale: Cldr.Locale.new!("zh")
-      "美元"
+      iex> Cldr.Currency.pluralize 12, :USD, locale: "zh"
+      {:ok, "美元"}
 
-      iex> Cldr.Currency.pluralize 12, :USD, locale: Cldr.Locale.new!("fr")
-      "dollars des États-Unis"
+      iex> Cldr.Currency.pluralize 12, :USD, locale: "fr"
+      {:ok, "dollars des États-Unis"}
 
-      iex> Cldr.Currency.pluralize 1, :USD, locale: Cldr.Locale.new!("fr")
-      "dollar des États-Unis"
+      iex> Cldr.Currency.pluralize 1, :USD, locale: "fr"
+      {:ok, "dollar des États-Unis"}
 
   """
   def pluralize(number, currency, options \\ []) do
@@ -135,13 +153,11 @@ defmodule Cldr.Currency do
 
     with \
       {:ok, currency_code} <- Cldr.validate_currency(currency),
-      {:ok, locale} <- Cldr.validate_locale(locale)
+      {:ok, locale} <- Cldr.validate_locale(locale),
+      {:ok, currency_data} <- currency_for_code(currency_code, locale)
     do
-      currency_data = for_code(currency_code, locale)
       counts = Map.get(currency_data, :count)
-      Cldr.Number.Cardinal.pluralize(number, locale, counts)
-    else
-      {:error, reason} -> {:error, reason}
+      {:ok, Cldr.Number.Cardinal.pluralize(number, locale, counts)}
     end
   end
 
@@ -161,11 +177,17 @@ defmodule Cldr.Currency do
   @doc """
   Returns a boolean indicating if the supplied currency code is known.
 
+  ## Options
+
   * `currency_code` is a `binary` or `atom` representing an ISO4217
-  currency code
+    currency code
 
   * `custom_currencies` is an optional list of custom currencies created by the
-  `Cldr.Currency.new/2` function
+    `Cldr.Currency.new/2` function
+
+  ## Returns
+
+  * `true` or `false`
 
   ## Examples
 
@@ -197,18 +219,29 @@ defmodule Cldr.Currency do
   custom currency code must start with an "X" followed by two alphabetic
   characters.
 
+  Note that since this function creates atoms but to a maximum of
+  26 * 26 == 676 since the format permits 2 alphabetic characters only.
+
+  ## Options
+
+  * `currency_code` is a `String.t` or and `atom` representing the new
+    currency code to be created
+
+  ## Returns
+
+  * `{:ok, currency_code}` or
+
+  * `{:error, {exception, message}}`
+
   ## Examples
 
       iex> Cldr.Currency.make_currency_code("xzz")
       {:ok, :XZZ}
 
       iex> Cldr.Currency.make_currency_code("aaa")
-      {:error,
-       "Invalid currency code \\"AAA\\".  Currency codes must start with 'X' followed by 2 alphabetic characters only."}
+      {:error, {Cldr.CurrencyCodeInvalid,
+       "Invalid currency code \\"AAA\\".  Currency codes must start with 'X' followed by 2 alphabetic characters only."}}
 
-  Note that since this function creates atoms, its important that this
-  function not be called with arbitrary user input since that risks
-  overflowing the atom table.
   """
   @valid_currency_code Regex.compile!("^X[A-Z]{2}$")
   @spec make_currency_code(binary | atom) :: {:ok, atom} | {:error, binary}
@@ -221,51 +254,50 @@ defmodule Cldr.Currency do
     if String.match?(currency_code, @valid_currency_code) do
       {:ok, String.to_atom(currency_code)}
     else
-      {:error, "Invalid currency code #{inspect currency_code}.  " <>
-        "Currency codes must start with 'X' followed by 2 alphabetic characters only."}
+      {:error, {Cldr.CurrencyCodeInvalid, "Invalid currency code #{inspect currency_code}.  " <>
+        "Currency codes must start with 'X' followed by 2 alphabetic characters only."}}
     end
   end
 
   @doc """
   Returns the currency metadata for the requested currency code.
 
-  * `currency_code` is a `binary` or `atom` representation of an ISO 4217 currency code.
+  ## Options
+
+  * `currency_code` is a `binary` or `atom` representation of an
+    ISO 4217 currency code.
 
   ## Examples
 
-      iex> Cldr.Currency.for_code("AUD")
-      %Cldr.Currency{cash_digits: 2, cash_rounding: 0, code: "AUD",
+      iex> Cldr.Currency.currency_for_code("AUD")
+      {:ok, %Cldr.Currency{cash_digits: 2, cash_rounding: 0, code: "AUD",
       count: %{one: "Australian dollar", other: "Australian dollars"},
       digits: 2, name: "Australian Dollar", narrow_symbol: "$",
-      rounding: 0, symbol: "A$", tender: true}
+      rounding: 0, symbol: "A$", tender: true}}
 
-      iex> Cldr.Currency.for_code("THB")
-      %Cldr.Currency{cash_digits: 2, cash_rounding: 0, code: "THB",
+      iex> Cldr.Currency.currency_for_code("THB")
+      {:ok, %Cldr.Currency{cash_digits: 2, cash_rounding: 0, code: "THB",
       count: %{one: "Thai baht", other: "Thai baht"}, digits: 2,
       name: "Thai Baht", narrow_symbol: "฿", rounding: 0, symbol: "THB",
-      tender: true}
+      tender: true}}
 
   """
-  @spec for_code(code, LanguageTag.t) :: %{}
-  def for_code(currency_code, locale \\ Cldr.get_current_locale()) do
+  @spec currency_for_code(code, LanguageTag.t) :: %{}
+  def currency_for_code(currency_code, locale \\ Cldr.get_current_locale()) do
     with \
       {:ok, code} <- Cldr.validate_currency(currency_code),
-      {:ok, locale} <- Cldr.validate_locale(locale)
+      {:ok, locale} <- Cldr.validate_locale(locale),
+      {:ok, currencies} <- currencies_for_locale(locale)
     do
-      locale
-      |> for_locale
-      |> Map.get(code)
-    else
-      {:error, reason} ->
-        {:error, reason}
+      {:ok, Map.get(currencies, code)}
     end
   end
 
   @doc """
   Returns the currency metadata for a locale.
   """
-  @spec for_locale(Locale.name | LanguageTag.t) :: Map.t
-  def for_locale(locale \\ Cldr.get_current_locale())
+  @spec currencies_for_locale(Locale.name | LanguageTag.t) :: Map.t
+  def currencies_for_locale(locale \\ Cldr.get_current_locale())
 
   for locale_name <- Cldr.Config.known_locale_names() do
     currencies =
@@ -273,21 +305,21 @@ defmodule Cldr.Currency do
       |> Cldr.Config.get_locale
       |> Map.get(:currencies)
       |> Enum.map(fn {k, v} -> {k, struct(@struct, v)} end)
-
-    def for_locale(%LanguageTag{cldr_locale_name: unquote(locale_name)}) do
-      unquote(Macro.escape(currencies))
       |> Enum.into(%{})
+
+    def currencies_for_locale(%LanguageTag{cldr_locale_name: unquote(locale_name)}) do
+      {:ok, unquote(Macro.escape(currencies))}
     end
   end
 
-  def for_locale(locale_name) when is_binary(locale_name) do
+  def currencies_for_locale(locale_name) when is_binary(locale_name) do
     case Locale.canonical_language_tag(locale_name) do
-      {:ok, locale} -> for_locale(locale)
+      {:ok, locale} -> currencies_for_locale(locale)
       {:error, reason} -> {:error, reason}
     end
   end
 
-  def for_locale(locale) do
+  def currencies_for_locale(locale) do
     {:error, Locale.locale_error(locale)}
   end
 end