+++ lib/number/symbol.ex
@@ -14,22 +14,51 @@ defmodule Cldr.Number.Symbol do
             :superscripting_exponent, :time_separator]
 
   @doc """
-  Returns a list of the number symbols for all number systems of a locale.
+  Returns a map of `Cldr.Number.Symbol.t` structs of the number symbols for each
+  of the number systems of a locale.
 
-  * `locale` is any locale known to `Cldr`.  See `/0/0s()`.
+  ## Options
+
+  * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
+    or a `Cldr.LanguageTag` struct returned by `Cldr.Locale.new!/1`.  The
+    default is `Cldr.get_current_locale/0`.
 
   ## Example:
 
-      iex> Symbol.number_symbols_for(Cldr.Locale.new("th"))
-      [latn: %{decimal: ".", exponential: "E", group: ",", infinity: "∞", list: ";",
-         minus_sign: "-", nan: "NaN", per_mille: "‰", percent_sign: "%",
-         plus_sign: "+", superscripting_exponent: "×", time_separator: ":"},
-       thai: %{decimal: ".", exponential: "E", group: ",", infinity: "∞", list: ";",
-         minus_sign: "-", nan: "NaN", per_mille: "‰", percent_sign: "%",
-         plus_sign: "+", superscripting_exponent: "×", time_separator: ":"}]
+      iex> Cldr.Number.Symbol.number_symbols_for("th")
+      {:ok, %{
+         latn: %Cldr.Number.Symbol{
+           decimal: ".",
+           exponential: "E",
+           group: ",",
+           infinity: "∞",
+           list: ";",
+           minus_sign: "-",
+           nan: "NaN",
+           per_mille: "‰",
+           percent_sign: "%",
+           plus_sign: "+",
+           superscripting_exponent: "×",
+           time_separator: ":"
+         },
+         thai: %Cldr.Number.Symbol{
+           decimal: ".",
+           exponential: "E",
+           group: ",",
+           infinity: "∞",
+           list: ";",
+           minus_sign: "-",
+           nan: "NaN",
+           per_mille: "‰",
+           percent_sign: "%",
+           plus_sign: "+",
+           superscripting_exponent: "×",
+           time_separator: ":"
+         }
+       }}
 
   """
-  @spec number_symbols_for(LanguageTag.t) :: Keyword.t
+  @spec number_symbols_for(LanguageTag.t | Locale.locale_name) :: Keyword.t
   def number_symbols_for(locale \\ Cldr.get_current_locale())
 
   for locale <- Cldr.Config.known_locale_names() do
@@ -37,17 +66,20 @@ defmodule Cldr.Number.Symbol do
       locale
       |> Cldr.Config.get_locale
       |> Map.get(:number_symbols)
+      |> Enum.map(fn
+          {k, nil} -> {k, nil}
+          {k, v} -> {k, struct(@struct, v)}
+         end)
+      |> Enum.into(%{})
 
     def number_symbols_for(%LanguageTag{cldr_locale_name: unquote(locale)}) do
-      symbols =
-        unquote(Macro.escape(symbols))
-        |> Enum.map(fn
-            {k, nil} -> {k, nil}
-            {k, v} -> {k, struct(__MODULE__, v)}
-           end)
-        |> Enum.into(%{})
+      {:ok, unquote(Macro.escape(symbols))}
+    end
+  end
 
-      {:ok, symbols}
+  def number_symbols_for(locale_name) when is_binary(locale_name) do
+    with {:ok, locale} <- Cldr.validate_locale(locale_name) do
+      number_symbols_for(locale)
     end
   end
 
@@ -58,49 +90,61 @@ defmodule Cldr.Number.Symbol do
   @doc """
   Returns the number sysbols for a specific locale and number system.
 
-  * `locale` is any locale known to `Cldr`.  See `/0/0s()`.
+  ## Options
 
-  * `number_system` which defaults to `:default` and is either:
+  * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
+    or a `Cldr.LanguageTag` struct returned by `Cldr.Locale.new!/1`.  The
+    default is `Cldr.get_current_locale/0`.
 
-    * an `atom` in which case it is interpreted as a `number system type`
-    in the given locale.  Typically this would be either `:default` or
-    `:native`. See `Cldr.Number.Format.format_types_for/1` for the number
-    system types available for a given `locale`.
-
-    * a `binary` in which case it is used to look up the number system
-    directly (for exmple `"latn"` which is common for western european
-    languages). See `Cldr.Number.Format.formats_for/1` for the
-    available formats for a `locale`.
+  * `number_system` is any number system name returned by
+    `Cldr.known_number_systems/0` or a number system type
+    returned by `Cldr.known_number_system_types/0`
 
   ## Example
 
-      iex> Cldr.Number.Symbol.number_symbols_for(Cldr.Locale.new("th"), "thai")
-      %{decimal: ".", exponential: "E", group: ",", infinity: "∞", list: ";",
-        minus_sign: "-", nan: "NaN", per_mille: "‰", percent_sign: "%",
-        plus_sign: "+", superscripting_exponent: "×", time_separator: ":"}
+      iex> Cldr.Number.Symbol.number_symbols_for("th", "thai")
+      {:ok, %Cldr.Number.Symbol{
+         decimal: ".",
+         exponential: "E",
+         group: ",",
+         infinity: "∞",
+         list: ";",
+         minus_sign: "-",
+         nan: "NaN",
+         per_mille: "‰",
+         percent_sign: "%",
+         plus_sign: "+",
+         superscripting_exponent: "×",
+         time_separator: ":"
+       }}
 
   """
-  @spec number_symbols_for(LanguageTag.t, atom | binary) ::
-    {:ok, Map.t} | {:no_symbols, String.t} | {:error, String.t}
+  @spec number_symbols_for(LanguageTag.t | Locale.locale_name, System.system_name) ::
+    {:ok, Map.t} | {:error, {Cldr.NoNumberSymbols, String.t}}
 
   def number_symbols_for(%LanguageTag{} = locale, number_system) do
-    with {:ok, system_name} <- Number.System.system_name_from(number_system, locale),
-         {:ok, symbols} <- number_symbols_for(locale)
+    with \
+      {:ok, system_name} <- Number.System.system_name_from(number_system, locale),
+      {:ok, symbols} <- number_symbols_for(locale)
     do
       symbols
       |> Map.get(system_name)
-      |> get_symbols_return(locale, number_system)
-    else
-      {:error, reason} -> {:error, reason}
+      |> symbols_return(locale, number_system)
     end
   end
 
-  defp get_symbols_return(nil, locale, number_system) do
-    {:no_symbols, "The locale #{inspect locale} does not have " <>
-        "any symbols for number system #{inspect number_system}"}
+  def number_symbols_for(locale_name, number_system) when is_binary(locale_name) do
+    with {:ok, locale} <- Cldr.validate_locale(locale_name) do
+      number_symbols_for(locale, number_system)
+    end
   end
 
-  defp get_symbols_return(symbols, _locale, _number_system) do
+  defp symbols_return(nil, locale, number_system) do
+    {:error, {Cldr.NoNumberSymbols, "The locale #{inspect locale} does not have " <>
+        "any symbols for number system #{inspect number_system}"}}
+  end
+
+  defp symbols_return(symbols, _locale, _number_system) do
     {:ok, symbols}
   end
 end